# perf: use non-forksafe RLock in context