--- a/ddtrace/context.py
+++ b/ddtrace/context.py
@@ -6,7 +6,6 @@ from typing import Text
 
 from .constants import ORIGIN_KEY
 from .constants import SAMPLING_PRIORITY_KEY
-from .internal import forksafe
 from .internal.compat import NumericType
 from .internal.logger import get_logger
 from .internal.utils.deprecation import deprecated
@@ -41,7 +40,7 @@ class Context(object):
         sampling_priority=None,  # type: Optional[float]
         meta=None,  # type: Optional[_MetaDictType]
         metrics=None,  # type: Optional[_MetricDictType]
-        lock=None,  # type: Optional[forksafe.ResetObject[threading.RLock]]
+        lock=None,  # type: Optional[threading.RLock]
     ):
         self._meta = meta if meta is not None else {}  # type: _MetaDictType
         self._metrics = metrics if metrics is not None else {}  # type: _MetricDictType
@@ -57,7 +56,7 @@ class Context(object):
         if lock is not None:
             self._lock = lock
         else:
-            self._lock = forksafe.RLock()
+            self._lock = threading.RLock()
 
     def _with_span(self, span):
         # type: (Span) -> Context