## Commit Message
{{title}}

Previously in #3036 we made a change to use a forksafe RLock
but did not re-measure the performance overhead of the change.

Creation of a `forksafe.RLock()` is 3 times slower than a
regular `threading.RLock`.

master   - Context(): Mean +- std dev: 3.24 us +- 0.34 us
v0.56.1  - Context(): Mean +- std dev: 754 ns +- 17 ns

Given that we do not _require_ a forksafe lock here we
should use a normal `threading.RLock` for performance reasons.