@P403n1x87 2021-12-03
That's a shame 😞 . I wonder if we could profile the forksafe lock and see if we could do anything to improve its performance?