@brettlangdon 2021-12-03
> That's a shame 😞 . I wonder if we could profile the forksafe lock and see if we could do anything to improve its performance?

We did a quick test on if the lock operations are also slowed down by about 2x. This is likely just showcasing the performance overhead from `wrapt.ObjectProxy`.

```
with threading.RLock(): Mean +- std dev: 326 ns +- 19 ns
with threading.Lock(): Mean +- std dev: 317 ns +- 28 ns
with forksafe.RLock(): Mean +- std dev: 619 ns +- 36 ns
with forksafe.Lock(): Mean +- std dev: 614 ns +- 34 n
```

<details>
<summary>Perf code</summary>

```python
import threading

import pyperf

from ddtrace.context import Context
from ddtrace.internal import forksafe


def add_args(args, cmd):
    cmd.rigorous = True


runner = pyperf.Runner(warmups=10, add_cmdline_args=add_args)


def run(lock):
    with lock:
        pass


runner.bench_func("with threading.RLock()", run, threading.RLock())
runner.bench_func("with threading.Lock()", run, threading.Lock())
runner.bench_func("with forksafe.RLock()", run, forksafe.RLock())
runner.bench_func("with forksafe.Lock()", run, forksafe.Lock())
```

</details>